+++ src/app/main/apps/protocolo-externo/protocolo-externo-detail/complementar/store/effects/documentos.effects.ts
@@ -20,7 +20,7 @@ import {environment} from 'environments/environment';
 import {AssinaturaService} from '@cdk/services/assinatura.service';
 import * as OperacoesActions from 'app/store/actions/operacoes.actions';
 import {ComponenteDigitalService} from '@cdk/services/componente-digital.service';
-import {DocumentosActionsAll} from '../actions';
+import {LoginService} from "../../../../../../auth/login/login.service";
 
 @Injectable()
 export class DocumentosEffects {
@@ -32,6 +32,7 @@ export class DocumentosEffects {
         private _documentoService: DocumentoService,
         private _componenteDigitalService: ComponenteDigitalService,
         private _assinaturaService: AssinaturaService,
+        private _loginService: LoginService,
         private _router: Router,
         private _store: Store<State>
     ) {
@@ -304,7 +305,10 @@ export class DocumentosEffects {
                         mergeMap((response: Assinatura) => [
                             new DocumentosActions.AssinaDocumentoEletronicamenteSuccess(action.payload.documentoId),
                             new AddData<Assinatura>({data: [response], schema: assinaturaSchema}),
-                            new DocumentosActions.GetDocumentos({'processoOrigem.id': `eq:${action.payload.processoId}`}),
+                            new DocumentosActions.GetDocumentos({
+                                'processoOrigem.id': `eq:${action.payload.processoId}`,
+                                'criadoPor.id': `eq:${this._loginService.getUserProfile().id}`
+                            }),
                             new OperacoesActions.Resultado({
                                 type: 'assinatura',
                                 content: `Assinatura id ${response.id} criada com sucesso!`,
@@ -327,7 +331,10 @@ export class DocumentosEffects {
                             .pipe(
                                 mergeMap(response => [
                                     new DocumentosActions.RemoveAssinaturaDocumentoSuccess(action.payload),
-                                    new DocumentosActions.GetDocumentos({'processoOrigem.id': `eq:${action.payload.processoId}`}),
+                                    new DocumentosActions.GetDocumentos({
+                                        'processoOrigem.id': `eq:${action.payload.processoId}`,
+                                        'criadoPor.id': `eq:${this._loginService.getUserProfile().id}`
+                                    }),
                                 ]),
                                 catchError((err, caught) => {
                                     console.log(err);